Update yo post persist hook for opportunity integration
Update yo post persist hook for opportunity integration

+++ hybris-sagia/hybris/bin/custom/sagia/sagiacore/src/com/sap/ibso/eservices/core/util/SagiaValueChainImageUtil.java
@@ -1,6 +1,5 @@
 package com.sap.ibso.eservices.core.util;
 
-
 import java.io.InputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -8,31 +7,28 @@ import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.HashSet;
 import java.util.List;
+import java.util.Objects;
 import java.util.Set;
 
 import org.apache.commons.codec.binary.Base64;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
 
 import com.investsaudi.portal.core.model.OpportunityProductModel;
-import com.sap.ibso.eservices.core.job.SagiaKeyStakeholdersImageGeneratorJob;
-
 import de.hybris.platform.catalog.CatalogVersionService;
 import de.hybris.platform.core.Registry;
 import de.hybris.platform.core.model.media.MediaModel;
-import de.hybris.platform.servicelayer.config.ConfigurationService;
 import de.hybris.platform.servicelayer.media.MediaService;
 import de.hybris.platform.servicelayer.model.ModelService;
 import de.hybris.platform.servicelayer.search.FlexibleSearchQuery;
 import de.hybris.platform.servicelayer.search.FlexibleSearchService;
 import de.hybris.platform.servicelayer.search.SearchResult;
 
-public class SagiaKeyStakeHoldersImageUtil {
+public class SagiaValueChainImageUtil {
 	
-	private static final Logger LOG = LoggerFactory.getLogger(SagiaKeyStakeHoldersImageUtil.class);
-	private static final String REAL_FILE_NAME = "keyStakeholders-";
+	private static final Logger LOG = LoggerFactory.getLogger(SagiaValueChainImageUtil.class);
+	private static final String REAL_FILE_NAME = "valueChain-";
 	private static final String SAGIA_PRODUCT_CATALOG = "sagiaProductCatalog";
 	private static final String CATALOG_VERSION_STAGED = "Staged";
 	private static final String PNG_MIME_TYPE = "application/png";
@@ -54,57 +50,61 @@ public class SagiaKeyStakeHoldersImageUtil {
 	
 	
 	
-	public void generateKeyStakeholdersImages(OpportunityProductModel opportunity)
+	public void generateValueChainImages(OpportunityProductModel opportunity)
 			throws Exception {
 		
 		Set<MediaModel> mediaList = new HashSet<>();
 		
-		for (MediaModel media : opportunity.getKeyStakeholders()) {
-			LOG.info("******************* Iterating over getKeyStakeholders *******************");
-			try {
-				if (StringUtils.isNotBlank(media.getEncodedString())) {
-					
-					if (isPng(media.getEncodedString())) {
-						LOG.info("The encoded string is a PNG image for media --> "+media.getCode());
-						byte[] decodedBytes = Base64.decodeBase64(media.getEncodedString());
-						LOG.info("******************* Before creating temp file *******************");
-						Path tempFile = Files.createTempFile(media.getCode(), PNG_FILE_TYPE);
-						try {
-							LOG.info("******************* Before writing temp file *******************");
-							Files.write(tempFile, decodedBytes);
+		if (Objects.nonNull(opportunity.getSupply())  &&  Objects.nonNull(opportunity.getSupply().getValueChainTexts())   &&  Objects.nonNull(opportunity.getSupply().getValueChainTexts().iterator().next().getAttachments())     ) {
+			for (MediaModel media : opportunity.getSupply().getValueChainTexts().iterator().next().getAttachments()) {
+				LOG.info("******************* Iterating over attachments *******************");
+				try {
+					if (StringUtils.isNotBlank(media.getEncodedString())) {
 
-							// Delete existing media if present
-							LOG.info("******************* Before deleting existing media *******************");
-							deleteExistingMedia(media.getCode());
+						if (isPng(media.getEncodedString())) {
+							LOG.info("The encoded string is a PNG image for media --> " + media.getCode());
+							byte[] decodedBytes = Base64.decodeBase64(media.getEncodedString());
+							LOG.info("******************* Before creating temp file *******************");
+							Path tempFile = Files.createTempFile(media.getCode(), PNG_FILE_TYPE);
+							try {
+								LOG.info("******************* Before writing temp file *******************");
+								Files.write(tempFile, decodedBytes);
 
-							// Create a new media model and save it to the media storage
-							LOG.info("******************* Before creating and saving new media *******************");
-							MediaModel newMedia = createAndSaveNewMedia(media, tempFile);
+								// Delete existing media if present
+								LOG.info("******************* Before deleting existing media *******************");
+								deleteExistingMedia(media.getCode());
 
-							// Import the file into media storage.
-							try (InputStream is = Files.newInputStream(tempFile)) {
-								LOG.info("******************* Before setting stream *******************");
-								mediaService.setStreamForMedia(newMedia, is);
+								// Create a new media model and save it to the media storage
+								LOG.info(
+										"******************* Before creating and saving new media *******************");
+								MediaModel newMedia = createAndSaveNewMedia(media, tempFile);
+
+								// Import the file into media storage.
+								try (InputStream is = Files.newInputStream(tempFile)) {
+									LOG.info("******************* Before setting stream *******************");
+									mediaService.setStreamForMedia(newMedia, is);
+								}
+
+								mediaList.add(newMedia);
+							} finally {
+								Files.delete(tempFile);
 							}
-
-							mediaList.add(newMedia);
-						} finally {
-							Files.delete(tempFile);
+							opportunity.getSupply().getValueChainTexts().iterator().next().setAttachments(mediaList);
+							modelService.save(opportunity.getSupply());
+							modelService.saveAll(opportunity.getSupply().getValueChainTexts());
+							LOG.info("Saving of decoded attachment Media for OpportunityProduct Supply ValueChain...");
+							modelService.save(opportunity);
 						}
-						opportunity.setKeyStakeholders(mediaList);
-						LOG.info("Saving of decoded keyStakeholders Media for OpportunityProduct...");
-						modelService.save(opportunity);
+
+						else {
+							LOG.info("The encoded string is NOT a PNG image for media --> " + media.getCode());
+						}
 					}
-					
-					else
-					{
-						LOG.info("The encoded string is NOT a PNG image for media --> "+media.getCode());
-					}
+				} catch (Exception e) {
+					LOG.error("Error while processing value chain media for OpportunityProduct with code " + opportunity.getCode(),
+							e);
 				}
-			} catch (Exception e) {
-				LOG.error("Error while processing keystakeholder media for OpportunityProduct with code " + opportunity.getCode(),
-						e);
-			}
+			} 
 		}
 		
 	}
@@ -114,7 +114,7 @@ public class SagiaKeyStakeHoldersImageUtil {
 
 		if (existingMediaList != null) {
 			for (MediaModel media : existingMediaList) {
-				LOG.info("Removing existing keyStakeholders media with code: {}", mediaCode);
+				LOG.info("Removing existing valuechain media with code: {}", mediaCode);
 				modelService.remove(media);
 			}
 		}
